add hideFullscreen component
