+++ src/components/post/hideFullscreen.tsx
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react';
+
+export default function HideFullscreen({ children }: { children: ReactNode }) {
+  return <div className='hide-fullscreen'>{children}</div>;
+}
