+++ src/components/post/postContentSection.tsx
@@ -1,28 +1,14 @@
-import PostSummarySection from '@/components/post/postSummarySection';
-import ToggleButtonItem from '@/components/post/toggleButtonItem';
+import { compoenents } from '@/lib/mdxComponents';
 import { MDXRemote } from 'next-mdx-remote/rsc';
-import { ReactNode } from 'react';
 import rehypePrettyCode from 'rehype-pretty-code';
 import rehypeSlug from 'rehype-slug';
-
-const ExternalLink = ({ children, ...props }: { children: ReactNode }) => {
-  return (
-    <a rel='noopener noreferrer' target='_blank' {...props}>
-      {children}
-    </a>
-  );
-};
 
 export default function PostContentSection({ content }: { content: string }) {
   return (
     <section className='prose post-content mb-20'>
       <MDXRemote
         source={content}
-        components={{
-          PostSummarySection,
-          ToggleButtonItem,
-          a: ExternalLink,
-        }}
+        components={compoenents}
         options={{
           mdxOptions: {
             rehypePlugins: [rehypePrettyCode, rehypeSlug],
