+++ src/posts/generator.mdx
@@ -4,26 +4,28 @@ date: '2025-08-24'
 tags: ['Kotlin', 'JavaScript', 'Python']
 ---
 
+<HideFullscreen>
 # 요약
-<PostSummarySection
-  summaryItems={[
-    {
-      title: '제너레이터란?',
-      contents: ['필요할 때마다 값을 하나씩 생성하는 이터레이터'],
-    },
-    {
-      title: '언제 사용될까?',
-      contents: [
-        '리스트를 한 방향으로 순회할 때 (ex. API 페이지네이션)',
-        '메모리 효율이 중요할 때 (ex. 파일 읽기, 무한 시퀀스)',
-      ],
-    },
-    {
-      title: '내부 원리는?',
-      contents: ['함수의 실행 상태를 저장한다.'],
-    },
-  ]}
-/>
+  <PostSummarySection
+    summaryItems={[
+      {
+        title: '제너레이터란?',
+        contents: ['필요할 때마다 값을 하나씩 생성하는 이터레이터'],
+      },
+      {
+        title: '언제 사용될까?',
+        contents: [
+          '리스트를 한 방향으로 순회할 때 (ex. API 페이지네이션)',
+          '메모리 효율이 중요할 때 (ex. 파일 읽기, 무한 시퀀스)',
+        ],
+      },
+      {
+        title: '내부 원리는?',
+        contents: ['함수의 실행 상태를 저장한다.'],
+      },
+    ]}
+  />
+</HideFullscreen>
 <br/>
 
 # 제너레이터, 그게 뭔데?
